+++ cypress/integration/2.1-graphql_testcases.spec.js
@@ -1,5 +1,4 @@
 import {
-  graphql,
   validateGetVersionResponse,
   validateGetViewListResponse,
   validateGetViewListsResponse,
@@ -19,6 +18,8 @@ import { loginViaCookies, updateRetry } from '../support/common/support.js'
 import { restAndGraphqlAPI } from '../support/outputvalidation.js'
 import { wipe } from '../support/api_testcase.js'
 import { syncCheckoutUICustom } from '../support/common/testcase.js'
+import { wishList } from '../support/app_list'
+import { graphql } from '../support/common/graphql_utils.js'
 
 describe('Graphql Testcase', () => {
   loginViaCookies({ storeFrontCookie: false })
@@ -28,37 +29,38 @@ describe('Graphql Testcase', () => {
   wipe()
 
   it('Mutation - addToList', updateRetry(3), () => {
-    graphql(addToList(payload), response => {
+    graphql(wishList, addToList(payload), response => {
       validateaddToListResponse(response)
       cy.setWishListItem('id', response.body.data.addToList)
     })
   })
 
   it('Query - Get Version', updateRetry(3), () => {
-    graphql(version(), validateGetVersionResponse)
+    graphql(wishList, version(), validateGetVersionResponse)
   })
 
   it('Query - Get checkList', updateRetry(3), () => {
-    graphql(checkList(), validateGetcheckListResponse)
+    graphql(wishList, checkList(), validateGetcheckListResponse)
   })
 
   syncCheckoutUICustom()
 
   it('Query - Get ViewList', updateRetry(3), () => {
-    graphql(viewList(payload), validateGetViewListResponse)
+    graphql(wishList, viewList(payload), validateGetViewListResponse)
   })
 
   it('Query - Get All ViewList', updateRetry(3), () => {
-    graphql(viewLists(), validateGetViewListsResponse)
+    graphql(wishList, viewLists(), validateGetViewListsResponse)
   })
 
   it('Query - Get listNames', updateRetry(3), () => {
-    graphql(listNames(), validateGetlistNamesResponse)
+    graphql(wishList, listNames(), validateGetlistNamesResponse)
   })
 
   it('Mutation - removeFromList', updateRetry(3), () => {
     cy.getWishListItem().then(wishlistid => {
       graphql(
+        wishList,
         removeFromList(wishlistid.id, payload),
         validateRemoveFromListResponse
       )