+++ cypress/support/app_list.js
@@ -1,0 +1,3 @@
+export default {
+  wishList: 'vtex.wish-list',
+}